+++ src/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter as Router} from 'react-router-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
 import './index.css';
 import App from './App';
 
-ReactDOM.render( 
-  <Router> <App/> </Router>, document.getElementById('root')
+ReactDOM.render(
+  <Router> <App /> </Router>, document.getElementById('root')
 );
 
 // If you want your app to work offline and load faster, you can change
