+++ src/routes/Routes.tsx
@@ -4,8 +4,8 @@ import { Switch } from 'react-router-dom';
 import Route from './Route';
 
 // pages
-import Login from '../login/login';
-import Pages from '../pages/pages';
+import Login from '../pages/login/login';
+import Pages from '../pages/home/pages';
 import PageNotFound from '../pages/page-not-found/page-not-found';
 
 export default function Routes() {
